Feature/hot fix after restart prject

See merge request Toever1503/cy-app-be!273

+++ src/main/java/cy/services/project/impl/ProjectServiceImpl.java
@@ -334,7 +334,10 @@ public class ProjectServiceImpl implements IProjectService {
                 }
             }
 
-            iFileRepository.delete(projectEntity.getAvatar());
+            if (projectEntity.getAvatar() != null){
+                iFileRepository.delete(projectEntity.getAvatar());
+            }
+
             if (projectModel.getAvatar() != null && !projectModel.getAvatar().isEmpty()) {
                 String urlAvatar = fileUploadProvider.uploadFile("avatar", projectModel.getAvatar());
                 FileEntity fileEntity = new FileEntity();
